Bme680.cpp bug fix in reading gas_valid_r and heat_stab_r bits
Now the par_g1, par_g2 and par_g3 are printed to system log too.

+++ src/Bme680.hpp
@@ -295,7 +295,7 @@ class Bme680 : public I2Chip
     void SetGasHeatResistor(uint8_t Profile, uint8_t R);
 
     /// Set gas heater temperature profile 0 - 9.
-    void SetGasHeatTemperature(uint8_t Profile, int Tamb, int T);
+    void SetGasHeatTemperature(uint8_t Profile, int8_t Tamb, uint16_t T);
 
     /// Set gas heater current bytes 0 - 9.
     void SetGasHeatCurrent(uint8_t Profile, uint8_t I);
